+++ apps/web/src/components/spinner.tsx
@@ -10,3 +10,16 @@ export function Spinner({ variant = 'default' }: SpinnerProps) {
   );
   return <div className={className} />;
 }
+
+export type SpinnerTimerProps = {
+  currentValue?: number;
+};
+
+export function SpinnerTimer({ currentValue }: SpinnerTimerProps) {
+  return (
+    <div className='relative flex items-center justify-center flex-col'>
+      <div className='absolute w-12 h-12 rounded-full border-2 border-white border-t-primary animate-spin'></div>
+      <span className='text-xl font-semibold text-slate-500'>{currentValue}</span>
+    </div>
+  );
+}