+++ packages/schemas/src/features/auth/login-user-credentials-schema.ts
@@ -1,7 +1,7 @@
 import z from 'zod';
 import { emailSchema } from './email-schema';
 
-export const loginCredentialsSchema = z.object({
+export const loginUserCredentialsSchema = z.object({
   email: emailSchema,
   password: z.string().trim(),
 });