+++ packages/schemas/src/features/auth/register-user-schema.ts
@@ -2,10 +2,10 @@ import z from 'zod';
 import { passwordSchema } from './password-schema';
 import { emailSchema } from './email-schema';
 
-export const userSchema = z
+export const registerUserCredentialsSchema = z
   .object({
     email: emailSchema,
-    password1: passwordSchema,
+    password: passwordSchema,
     password2: z.string(),
   })
-  .refine(user => user.password1 === user.password2, { error: 'Passwords do not match!' });
+  .refine(user => user.password === user.password2, { error: 'Passwords do not match!' });